refactor: extract useDashboardData from Dashboard
Extract data fetching functions and their useEffects into useDashboardData
hook (555 lines). Removes ~425 lines from Dashboard.js (3314 -> 2889).

Hook contains: fetchApiKey, fetchData (~200 lines NWC+Blink transaction
fetcher), fetchWallets, fetchVoucherWalletBalance, getCapacityColor,
plus useEffects for: fetch-on-user-change, display-currency-set,
refresh-on-view-switch, voucher-balance-on-view, poll-voucher-amount,
fetch-on-overlay-open, refresh-on-wallet-change, fetch-wallets-on-apiKey,
refresh-on-payment-received.

Also removes unused getApiUrl, getAllValidDomains imports from Dashboard.
